feat: Refactor data layer to improve event and sport fetching with new repository methods and local data sources
feat: Enhance event fetching by adding favorite filter to getEventsBySport method

+++ core/data/src/main/java/g/sig/core/data/remote/enitites/EventApiDto.kt
@@ -20,7 +20,7 @@ data class EventApiDto(
             id = id,
             sportId = sportId,
             name = name,
-            startTime = startTime,
+            startTime = startTime.toLong(),
             isFavorite = false,
         )
 }
